requirements.txt
Django==5.1.3
Pillow==10.1.0
qrcode==7.4.2
python-dotenv==1.0.0
pyzbar==0.1.9
